Create README.md
God fearing, and doesn't work on Saturdays. Hardworking, agile and have excellent communication skills with high proficiency in the use of English both written and spoken. I don't not work well with crowds so I prefer to work remotely but at the same time I work well with teams of software developers.